last changes  today

+++ lib/models/payment_model.dart
@@ -2,7 +2,7 @@ class PaymentModel {
   String? paymentNo;
   String? controlNo;
   double? amount;
-  int? units;
+  double? units;
   String? date;
 
   PaymentModel({
